ADD: duplicate detection based on solr dismax query parser FIX: prices in timbirichi

+++ scraper/pipelines.py
@@ -10,6 +10,8 @@ import logging
 from django.contrib.postgres.search import TrigramSimilarity
 from django.db.models import Q
 from django.utils import timezone
+from haystack.inputs import Raw
+from haystack.query import SearchQuerySet
 
 from ads.models import Ad
 from comohay import settings
@@ -19,27 +21,38 @@ class RemoveDuplicatedAdPipeline(object):
 
     def process_item(self, item, spider):
 
-        if spider.name not in ['hogarencuba', 'updater']: # hogarencuba has no duplicates and updater only updates
+        if spider.name not in ['hogarencuba', 'updater'] and item['description'].strip() != '':  # hogarencuba has no duplicates and updater only updates
+            sqs = SearchQuerySet()
+            clean_desc = sqs.query.clean(item['description'])
+            similarity = int(settings.DESCRIPTION_SIMILARITY * 100)
+            max_len = len(item['description']) + int(len(item['description']) * settings.DESCRIPTION_LENGTH_DIFF)
+            ids_values = sqs.filter(
+                content=Raw("(description_length:[0 TO {}]) AND {{!dismax qf=description mm={}%}}{}".format(max_len, similarity, clean_desc,))
+            ).values_list('id')
+            ids = list(map(lambda x: x[0].split('.')[-1], ids_values))
+
             try:
                 # Remove duplicated ads from same contact
-                a = Q(contact_email=item['contact_email']) & Q(contact_email__isnull=False) & ~Q(contact_email__exact='')
-                b = Q(contact_phone=item['contact_phone']) & Q(contact_phone__isnull=False) & ~Q(contact_phone__exact='')
-                c = Q(external_contact_id=item['external_contact_id']) \
+
+                a = Q(id__in=ids)
+                b = Q(contact_email=item['contact_email']) & Q(contact_email__isnull=False) & ~Q(
+                    contact_email__exact='')
+                c = Q(contact_phone=item['contact_phone']) & Q(contact_phone__isnull=False) & ~Q(
+                    contact_phone__exact='')
+                d = Q(external_contact_id=item['external_contact_id']) \
                     & Q(external_contact_id__isnull=False) \
                     & ~Q(external_contact_id__exact='') \
                     & Q(external_source=item['external_source'])
-                Ad.objects.annotate(
-                    desc_similarity=TrigramSimilarity('description', item['description'])
-                ).filter(
-                    a | b | c
-                ).filter(
-                    desc_similarity__gt=settings.DESCRIPTION_SIMILARITY,
+
+                Ad.objects.filter(
+                    a & (b | c | d)
                 ).exclude(
                     external_source=item['external_source'],
                     external_id=item['external_id']
                 ).delete()
+
             except Exception as e:
-                logging.error("Error removing duplicated items: "+str(e))
+                logging.error("Error removing duplicated items: " + str(e))
         return item
